+++ CHANGELOG.md
@@ -6,6 +6,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unrleased]
 
+## 1.1.0
+
+### Added
+
+- `checkImage()` function to remove loading spinner after image is loaded.
+- Comments to background.js
+- Dropdown for meme names.
+
+### Changed
+
+- Progress bar to loading spinner.
+- Make meme button to be larger and span the element.
+- Image output to the bottom of the page.
+
+### Fixed
+
+- Broken memes when URL contains a `-`
+- URI encoding for special characters
+
+### Removed
+
+- Text about copy/paste into a chat.
+- List of meme names
+
 ## 1.0.1
 
 ### Changed